Correctly handle multiple set-cookie headers in response
This commit changes how set-cookies headers with multiple values
are returned. Previously, the code looks for Array types in the
headers, but multiple values are actually stored as a comma
delimited list. Following the adapter-netlify approach, the
set-cookie-parser package is used to split the set-cookie header
while all other headers are left unchanged.