Test processing of set-cookie headers in SvelteKit response
This commit adds a vitest suite and a test harness to allow
testing of serverless.js. The files serverless.js and shim.js are
copied to a folder where index.js and manifest.js are stored as stubs
in the parent folder. They can then be used within the Vitest testing
framework.

The tests here check that set-cookie headers are converted to
multiValueHeaders in the lambda response, while any other header
remains in the header field, being comma separated if multiple values
are set.